Field now can run more kernels in parallel. The current time is now computed directly removing roundoff error.

+++ inc/simulation_state.h
@@ -7,7 +7,7 @@ class SimulationState
    ~SimulationState(){};
    static SimulationState & getRef();
 
-   float simTime;
+   double simTime;
    unsigned int iterationNum;
    unsigned int numEleHot;
    unsigned int numEleCold;
